otp send mail and update the user password

+++ src/main/java/com/Power_gym/backend/DTO/UserScheduleDetailDTO.java
@@ -14,14 +14,14 @@ public class UserScheduleDetailDTO {
     private Integer scheduleDetailID;
     private Integer age;
     private String gender;
-    private Integer height;
+    private Double height;
     private Integer experience;
-    private Integer weight;
+    private Double weight;
     private Double bmi;
     private Integer workoutTime;
     private String fitnessGoal;
     private String createTime;
     private Integer isActive;
     private Schedule schedule;
-    private Integer userId;
+    private Long userId;
 }
